# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "description": "账号名唯一；注册成功直接返回令牌，无需再登录一次。\n\n---\n\n**鉴权**：无需令牌，公开接口。",
        "operationId": "auth.register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AuthTokenDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "注册账号",
        "tags": [
          "公共接口"
        ],
        "x-olua-audience": "public",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AuthTokenDto"
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "账号不存在与密码错误返回同一句提示（不暴露账号是否存在）。\n\n**失败限流**：同一用户名连续失败 5 次、或同一 IP 在 10 分钟内失败 20 次，会被临时锁定并返回 429 / 10005（阈值见服务端 `.env` 的 `LOGIN_*`）。\n\n---\n\n**鉴权**：无需令牌，公开接口。",
        "operationId": "auth.login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AuthTokenDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "账号登录",
        "tags": [
          "公共接口"
        ],
        "x-olua-audience": "public",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AuthTokenDto"
        }
      }
    },
    "/auth/me": {
      "get": {
        "description": "用令牌里的账号 id 查，不接受任何入参指定账号。\n\n---\n\n**鉴权**：客户端令牌（`player`），只能操作令牌所属账号自己的数据。",
        "operationId": "auth.me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AccountDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "player": []
          }
        ],
        "summary": "获取当前登录账号",
        "tags": [
          "客户端"
        ],
        "x-olua-audience": "player",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AccountDto"
        }
      }
    },
...
    "/admin/auth/register": {
      "post": {
        "description": "服务端配置了 `ADMIN_REGISTER_CODE` 时，registerCode 必须一致，否则拒绝注册。**第一个**注册的管理员自动成为超级管理员，之后注册的一律是普通管理员。\n\n---\n\n**鉴权**：无需令牌，公开接口。",
        "operationId": "adminAuth.register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminTokenDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "注册管理员",
        "tags": [
          "公共接口"
        ],
        "x-olua-audience": "public",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AdminTokenDto"
        }
      }
    },
    "/admin/auth/login": {
      "post": {
        "description": "成功后把返回的 token 放到 `Authorization: Bearer <token>`。\n\n**失败限流**：同一用户名连续失败 5 次、或同一 IP 在 10 分钟内失败 20 次，会被临时锁定并返回 429 / 30009（阈值见服务端 `.env` 的 `LOGIN_*`）。\n\n---\n\n**鉴权**：无需令牌，公开接口。",
        "operationId": "adminAuth.login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminTokenDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "管理员登录",
        "tags": [
          "公共接口"
        ],
        "x-olua-audience": "public",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AdminTokenDto"
        }
      }
    },
    "/admin/auth/me": {
      "get": {
        "description": "返回当前管理员的角色与**权限点清单** —— 管理端据此显示 / 隐藏菜单与按钮（服务端仍会独立校验，前端隐藏只是体验）。\n\n---\n\n**鉴权**：管理端令牌（`admin`）；本接口只操作当前登录管理员自身，不需要额外权限点。",
        "operationId": "adminAuth.me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "当前管理员",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "获取当前登录管理员",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AdminDto"
        },
        "x-olua-permissions": []
      }
    },
    "/admin/auth/password": {
      "patch": {
        "description": "任何已登录管理员都能改**自己**的密码，需要带上原密码（光有令牌不该能改口令）。\n\n改完直接返回**新的令牌**：库里令牌版本号已 +1，此前签发的令牌全部作废 —— 管理端应当用返回的 token 覆盖本地那一份，否则会被自己的系统踢回登录页。\n\n---\n\n**鉴权**：管理端令牌（`admin`）；本接口只操作当前登录管理员自身，不需要额外权限点。",
        "operationId": "adminAuth.changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeAdminPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "改密成功（含新令牌）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminTokenDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "修改自己的密码",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AdminTokenDto"
        },
        "x-olua-permissions": []
      }
    },
...
    "/admin/admins/{id}/password": {
      "patch": {
        "description": "只有超级管理员能做（`admin:manage`）—— 管理员之间不能互相改密码。\n\n**副作用**：对方此前签发的令牌立即作废，他会被踢回登录页，需要用新密码重新登录。不触发「最后一个超管」保护：改密码不会让人失去登录能力，改的也不是身份与状态。管理员想改自己的密码应当走 `PATCH /admin/auth/password`（需要原密码）。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`admin:manage`（管理管理员；角色：超级管理员）",
        "operationId": "adminAdmin.resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "管理员 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetAdminPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "重置结果（不回显密码）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ResetPasswordResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "重置某个管理员的密码",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/ResetPasswordResultDto"
        },
        "x-olua-permissions": [
          "admin:manage"
        ]
      }
    }
  },
  "info": {
    "title": "olua 游戏服务端 API",
    "description": "olua 客户端与管理端共用的服务端接口。\n\n**接口分三组**（左侧按组展示，顺序即权限层级）：\n1. **公共接口** —— 无需令牌（注册 / 登录 / 健康检查）\n2. **客户端** —— 需要 `player` 令牌（`/api/auth/login` 获取），只能操作自己账号的数据\n3. **管理端** —— 需要 `admin` 令牌（`/api/admin/auth/login` 获取），且每个接口还要求权限点\n\n**统一响应包裹**：所有接口（成功与失败）都是 `{ code, message, data, timestamp }`；\n`code = 0` 表示成功，其余见服务端 `common/constants/biz-code.ts`。\n\n**权限**：管理员分三级 —— 超级管理员 / 管理员 / 只读观察员，\n每个管理端接口的说明里都写了「所需权限」，由 `common/constants/permission.ts` 统一维护。\n请求头一律是 `Authorization: Bearer <token>`。\n\n---\n\n**这份文档是三端接口文件的唯一来源**：客户端 `assets/ui/utils/net/` 与管理端 `admin/src/api/`\n里的路径、类型、接口方法都由 `tools/gen-api.cjs` 从本文件生成，\n所以**改接口先改服务端**，再执行 `cd server && npm run gen:api` 重新生成（禁止手改生成物）。\n生成器读取的机器可读约定：\n- 每个接口的 `operationId` 形如 `<模块>.<方法>`（如 `role.save`），决定生成到哪个模块的哪个方法；\n- `x-olua-audience` 标明调用方（public / player / admin）；\n- `x-olua-permissions` 是管理端接口要求的权限点；\n- `x-olua-data-schema` 是响应里 `data` 的真实类型（没有该扩展 = 该接口无业务数据）。",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "player": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "客户端令牌（/api/auth/login 获取）"
      },
      "admin": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "管理端令牌（/api/admin/auth/login 获取）"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "游戏账号：3~20 位字母 / 数字 / 下划线",
            "example": "player001"
          },
          "password": {
            "type": "string",
            "description": "密码：6~32 位",
            "example": "123456"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "游戏账号",
            "example": "player001"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "123456"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
      "AdminRegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "管理员账号：3~20 位字母 / 数字 / 下划线",
            "example": "gm001"
          },
          "password": {
            "type": "string",
            "description": "密码：6~32 位",
            "example": "admin123"
          },
          "registerCode": {
            "type": "string",
            "description": "注册码（服务端配置了 ADMIN_REGISTER_CODE 时必填）",
            "example": "olua-admin-2026"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "AdminLoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "管理员账号",
            "example": "gm001"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "admin123"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "ChangeAdminPasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "当前密码"
          },
          "newPassword": {
            "type": "string",
            "description": "新密码（6~32 位）"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
...
      "ResetAdminPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "新密码（6~32 位）",
            "example": "new-pass-123"
          }
        },
        "required": [
          "password"
        ]
      },
      "ApiEnvelopeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "业务码，0 = 成功",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "example": "ok"
          },
          "data": {
            "type": "object",
            "description": "业务数据（无数据时为 null）",
            "additionalProperties": true,
            "nullable": true
          },
          "timestamp": {
            "type": "number",
            "description": "服务端时间戳（毫秒）",
            "example": 1760000000000
          },
          "path": {
            "type": "string",
            "description": "出错时的请求路径",
            "example": "/api/auth/login"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "timestamp"
        ]
      },
...
      "AccountDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "账号 id",
            "example": "6f1c2b1e-3a5d-4c9e-9a1f-2b3c4d5e6f70"
          },
          "username": {
            "type": "string",
            "description": "账号名",
            "example": "player001"
          },
          "status": {
            "type": "string",
            "description": "状态：active 正常 / disabled 封禁",
            "example": "active"
          },
          "onlineRoleId": {
            "type": "string",
            "description": "当前选中的角色 id（未选角色时为 null）",
            "nullable": true,
            "example": null
          },
          "roleCount": {
            "type": "number",
            "description": "角色数量（列表接口返回）",
            "example": 2
          },
          "createdAt": {
            "type": "number",
            "description": "注册时间（毫秒）",
            "example": 1760000000000
          },
          "updatedAt": {
            "type": "number",
            "description": "最近更新时间（毫秒）",
            "example": 1760000000000
          },
          "lastLoginAt": {
            "type": "number",
            "description": "最近登录时间（毫秒，从未登录为 null）",
            "nullable": true,
            "example": 1760000000000
          }
        },
        "required": [
          "id",
          "username",
          "status",
          "onlineRoleId",
          "createdAt",
          "updatedAt",
          "lastLoginAt"
        ]
      },
      "AuthTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "访问令牌（放 Authorization: Bearer <token>）"
          },
          "expiresIn": {
            "type": "string",
            "description": "令牌有效期（与 JWT_EXPIRES_IN 一致）",
            "example": "7d"
          },
          "account": {
            "description": "账号信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/AccountDto"
              }
            ]
          }
        },
        "required": [
          "token",
          "expiresIn",
          "account"
        ]
      },
...
      "AdminDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "管理员 id"
          },
          "username": {
            "type": "string",
            "description": "管理员账号名"
          },
          "status": {
            "type": "string",
            "description": "状态：active 正常 / disabled 停用"
          },
          "role": {
            "type": "string",
            "description": "管理员角色：super_admin 超级管理员 / admin 管理员 / viewer 只读观察员",
            "enum": [
              "super_admin",
              "admin",
              "viewer"
            ]
          },
          "roleLabel": {
            "type": "string",
            "description": "角色中文名（界面直接展示）"
          },
          "permissions": {
            "type": "array",
            "description": "该角色拥有的权限点（管理端据此显示 / 隐藏按钮；服务端仍会独立校验）",
            "items": {
              "type": "string"
            },
            "example": [
              "stats:read",
              "account:read"
            ]
          },
          "createdAt": {
            "type": "number",
            "description": "创建时间（毫秒）"
          },
          "lastLoginAt": {
            "type": "number",
            "description": "最近登录时间（毫秒，从未登录为 null）",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "status",
          "role",
          "roleLabel",
          "permissions",
          "createdAt",
          "lastLoginAt"
        ]
      },
      "AdminTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "访问令牌（放 Authorization: Bearer <token>）"
          },
          "expiresIn": {
            "type": "string",
            "description": "令牌有效期",
            "example": "7d"
          },
          "admin": {
            "description": "管理员信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/AdminDto"
              }
            ]
          }
        },
        "required": [
          "token",
          "expiresIn",
          "admin"
        ]
      },
...
      "ResetPasswordResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "被重置的对象 id"
          },
          "name": {
            "type": "string",
            "description": "被重置的对象名（账号名 / 管理员名，界面直接展示）"
          },
          "revokedTokens": {
            "type": "boolean",
            "description": "提示：口令变更后，该对象此前签发的令牌**已全部作废**，需要用它重新登录",
            "example": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
    }
  }
}